Move file to Memory/ and edit EffectAction to consider deleting effect action
